Removed several broken playlists
This project depends on information hosted by third parties on their servers.

Under the conditions of the Russian special operation (which continues February 24, 2022 to the present), some playlists have mutated into geopolitical and anti-Russian slogans.

I see this as totally unacceptable. The open source world is not the place for this kind of debate. Relevant projects cannot be used to impose someone else's opinion.

At the same time, in the current list of playlists, both Ukrainian and Russian channels are on an equal footing. The user must have their own choice in consuming content and seeking the truth.